warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@11509 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Sessions/Src/Main.cpp
@@ -30,7 +30,7 @@ bool g_hghostw;
 HWND hClistControl;
 
 int g_ses_limit;
-size_t g_ses_count;
+int g_ses_count;
 bool g_bExclHidden;
 bool g_bWarnOnHidden;
 bool g_bOtherWarnings;
